+++ data/solutions/ctrl_crypto_001.py
@@ -1,0 +1,6 @@
+hash_value = 0x00000ABC
+target = 0x0000FFFF
+if hash_value < target:
+    valid = True
+else:
+    valid = False