+++ data/solutions/ctrl_hack_001.py
@@ -1,0 +1,7 @@
+clearance = 3
+if clearance == 3:
+    permission = "admin"
+elif clearance == 2:
+    permission = "operator"
+else:
+    permission = "guest"