+++ data/solutions/func_hack_001.py
@@ -1,0 +1,2 @@
+def decode_char(code):
+    return chr(code)