+++ data/solutions/func_sci_001.py
@@ -1,0 +1,2 @@
+def kinetic_energy(mass, velocity):
+    return 0.5 * mass * velocity**2